Fix compile error for Unity build

+++ PythonClient/airsim/client.py
@@ -116,7 +116,7 @@ class VehicleClient:
         Continue (or resume if paused) the simulation for the specified number of frames, after which the simulation will be paused.
 
         Args:
-            frames (float): Frames to run the simulation for
+            frames (int): Frames to run the simulation for
         """
         self.client.call('simContinueForFrames', frames)
